Made type changes to Customer and customerDto;

+++ CustomerTransaction.SharedEntities/Models/Customer.cs
@@ -8,7 +8,7 @@ namespace CustomerTransaction.SharedEntities.Models
         public Int64 CustomerId { get; set; }
         public String Name { get; set; }
         public String Email { get; set; }
-        public Int32 Mobile { get; set; }
+        public Int64 Mobile { get; set; }
         public IEnumerable<Transaction> Transactions { get; set; }
     }
 }
